Move MicroservicesServerImpl to internal packages
