+++ core/src/main/java/org/wso2/msf4j/internal/MicroservicesServerSC.java
@@ -34,7 +34,6 @@ import org.wso2.msf4j.Interceptor;
 import org.wso2.msf4j.Microservice;
 import org.wso2.msf4j.MicroservicesRegistry;
 import org.wso2.msf4j.MicroservicesServer;
-import org.wso2.msf4j.MicroservicesServerImpl;
 import org.wso2.msf4j.SessionManager;
 import org.wso2.msf4j.SwaggerService;
 import org.wso2.msf4j.exception.OSGiDeclarativeServiceException;
